VSCPP Compatibility fixes

+++ libewf/libewf_segment_file.c
@@ -1519,7 +1519,7 @@ ssize_t libewf_segment_file_read_table2_section(
 	{
 		segment_file->flags |= LIBEWF_SEGMENT_FILE_FLAG_IS_CORRUPTED;
 
-		number_of_entries = chunk_group_number_of_entries;
+		number_of_entries = (uint32_t) chunk_group_number_of_entries;
 	}
 	else if( number_of_entries_mismatch != 0 )
 	{
@@ -1528,7 +1528,7 @@ ssize_t libewf_segment_file_read_table2_section(
 		{
 			/* If the number of entries in the table section are not 0 prefer the table section
 			 */
-			number_of_entries = chunk_group_number_of_entries;
+			number_of_entries = (uint32_t) chunk_group_number_of_entries;
 		}
 		else
 		{
